+++ src/frontend/components/chrome/footer/i18n/en.ts
@@ -5,7 +5,6 @@ export default defineMessages({
     nav: {
       contact: "Contact",
       hours: "Opening hours",
-      news: "News",
       offer: "What we offer",
       pricing: "Prices",
     },